tasklist

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+       
+        
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        }
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        return view('welcome',$data);
         //
     }
 
@@ -50,11 +60,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:191',
         ]);
-        
-        $task = new Task;
-        $task->status = $request->status; 
-        $task->content = $request->content;
-        $task->save();
+
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status'=> $request->status,
+        ]);
 
         return redirect('/');
         //
@@ -69,11 +79,13 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
-        //
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        
+        return redirect('/');
     }
 
     /**
@@ -85,11 +97,12 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
-        //
+        if (\Auth::id() === $tasks->user_id){
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -106,10 +119,11 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         $task = Task::find($id);
-        $task->status = $request->status; 
-        $task->content = $request->content;
-        $task->save();
-
+        if (\Auth::id() === $tasks->user_id){
+            $task->status = $request->status; 
+            $task->content = $request->content;
+            $task->save();
+        }
         return redirect('/');
         //
     }
@@ -123,8 +137,9 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
-
+        if (\Auth::id() === $tasks->user_id){
+            $task->delete();
+        }
         return redirect('/');
         //
     }
